update segment routing config and accomodate area_id and area_type in area_config
Summary:
Currently, we allocate a unique node segment label per device. Going forward, each device will be a part of multiple areas and we would like to support segment routing in such a multi-area network.

In order to support MPLS routing in multi-area network, we would like to allocate a unique node segment label per area on a given device.

- First we will have one area (specifically mesh-area).
- Going forward, a device will be a part of pod-area, mesh-area and spine-area.

Enhancing the existing `OpenrConfig.thrift` to accommodate the new requirement.
- Added checks in config parsing

Deployment wise, although we have rolled out the thrift structure but this change will not break anything as we have not made changes to parse node segment label config.

In addition to `area_id`, adding a field called `area_type` so that searching through config can be done based on type (instead of `area_id`).
For example, if there is one mesh, openr does not need to use exact name.

Reviewed By: xiangxu1121

Differential Revision: D28547646

fbshipit-source-id: a76ba21bcb85f35fe9be7e192e1809c2e22a1999