Fib - Filter MPLS next-hops before storing or programming
Summary:
We filter MPLS next-hops as soon as the update is received from Decision. This
means subsequent programming or re-programming of route on failure doesn't need
to handle this case.
A lot of code gets simplified as a result.

Minor change is to create a separate function for processing static route
update.

Reviewed By: cfdream

Differential Revision: D29014435

fbshipit-source-id: c12736ed2b5ee827e77e96c8476ddc5340b8e86d