Add nexthop filtering logic in RibMplsEntry
Summary:
For MPLS nexthops, we need to filter next-hops to a unique action type. It is
possible for control plane to compute next-hops that are mix of both PHP and
SWAP. But HW can only program a unique action per NHG.

We do so as of now via `Util.selectMplsNextHops` function. This diff now
implements the same logic in RibMplsEntry. Hoping to use this in subsequent diff
that'll greatly simplify some of the logic.

Reviewed By: cfdream

Differential Revision: D29012344

fbshipit-source-id: b008b76aa48386199404c78fe8ba5408c1da57e3