Fib - Incremental programming of failed routes
Summary:
This diff implements incremental route programming on failure. Before this diff
on encountering failure we programm all the routes again. With a specific
exception supported from PlatformAgent about route programming failures, we can
now understand which routes failed to program.

== Error Handling
There are two types of exception we may receive. `PlatformFibUpdateError` and
`std::exception`. The former is specific in terms of what went wrong during
programming while later one is not so specific. For later one we assume that
everything failed to program.

We use explicit `try-catch` for every route programming API. This helps in few
ways
- Processing error conditions
- Proceed with deletion of route if programming fails aka best effort
- Whole logic (aka logging, fib update publication) is same in dryrun state now.
  Just that we skip programming

== State Management
Below are the major changes in the way we represent Route State
- Removed `FIB_ERROR` event, as it is not longer needed for state transition.
- The `State` remain same on `PlatformFibUpdateError`. And we just do
  incremental route programming
- We record failed updated as dirty prefixes and labels
- We use same `retryRoutesTimer` for retry programming of failed routes

== FIB Update Publication
- On successful first sync we publish update with type `FULL_SYNC`. All
  subsequent update will have type `INCREMENTAL`
- On subsequent full-sync we only publish dirty-state or empty update (for ease
  of testing)
- On receipt of route from Decision, we program it. On failure we update

Reviewed By: cfdream

Differential Revision: D29081999

fbshipit-source-id: ac539bf7ee636451a7fb53864d328470a92db6b5