Remove redundant enum type to avoid unexpected state
Summary:
We should only define enumerations that succintly represent a state and are
intended to be used in code. In this case, here we remove type `DEFAULT` as it
is not intended to be used. Code doesn't handle the case if type is enum and it
introduces chances for data curruption.

We set default value of enum to be `INCREMENTAL`, a behavior well expected
throughout the code. Only Fib.cpp will overwrite INCREMENTAL type to FULL_SYNC
for first publication it exports to its readers. This is well handled in UTs.

Reviewed By: saumitrp

Differential Revision: D29341509

fbshipit-source-id: b668d2bd1ce45b747f4324594dd5ce08c288839c