Add KvRequestQueue to send KvStore requests to update its keys
Summary:
Part of KvStoreClientInternal deprecation effort.

Add KeyValueRequest, which is the union of all types of requests that can be sent: SetKeyValueRequest, PersistKeyValueRequest, EraseKeyValueRequest (more to come later).

This request can be sent from other modules (e.g. PrefixManager, LinkMonitor) to update the keys in KvStore.

Update unit tests that use this inter-module communication to have this queue.

Queue is only set up, not yet used.

Reviewed By: xiangxu1121

Differential Revision: D29387416

fbshipit-source-id: 9cab49bbfa75209fc3f6a8476daaf4f15d8982ef