Spark - Ignore neighbor's v4-addr if local node has v4-over-v6 enabled
Summary:
If v4-over-v6 is enabled then the local node is never going to use v4 interface
address. Hence we must not validate it.

After whole network has v4-over-v6 enabled, the next step would be to disable
v4. When we disable v4 the previous implementation in Spark ignores it.

Differential Revision: D30051962

fbshipit-source-id: 6cd2be92cbdb60a706a6859dcd976393650af500