(7/N): Leverage getTry() with folly::Semifuture to make
Summary:
This is the continuous efforts with fiber task adoption inside LinkMonitor to make sure evb won't stuck and get killed by `Watchdog`.

Even with fiber task usage, it is still possible to be stuck when `get()` call is blocked for this fiber task particularly.

This diff leverages `getTry(HighResDuration)` inside folly to help with, which is similar to add a client-side timeout in server-client mode.

Reviewed By: saifhhasan

Differential Revision: D30145730

fbshipit-source-id: 6650e7c5022bd1acbb6707e7dc3b62708e2e9aba