Add unit test for static node segment label allocation
Summary:
Updated unit tests which test node segment label distribution among peers when multiple areas are configured.

The changes are not for kinds of label allocation - dynamic and static.

Each AdjDb is expected to have node segment label entries for all nodes in a given area (observed in a given openr instance).

Reviewed By: saifhhasan

Differential Revision: D30093098

fbshipit-source-id: 8540a09c2d9c0c67a55b178aefb0f1f61a1be199