Support Per Area Forwarding Algorithm
Summary:
This diff refactors several SpfSolver path computation methods to operate on one are at a time. These are the following:
selectBestPathsSpf()
selectBestPathsKsp2()
getNextHopsWithMetric()
getNextHopsThrift()

In addition, selectBestPathsSpf() and selectBestPathsKsp2() now return a set of next-hops (previously they returned RibUnicastEntry). The next-hops for each area are merged together in the calling function createRouteForPrefix()

createRouteForPrefix() walks the area path computation rules and calls either selectBestPathsSpf() or selectBestPathsKsp2() based on the area's forwarding algorithm. Note that if the default route computation rules are used (see code in SpfSolver::getRouteComputationRules()), all areas that have selected routes are used.

This diff also fixes the bug in T71111090. We now only consider the shortest path in the area the route was received in. Some of the expected results of DecisionTest UTs had to be updated for the new behavior.

Reviewed By: saifhhasan, cfdream

Differential Revision: D29970978

fbshipit-source-id: a5416a237ce55271c822b70b8dbe760849182f33