WatchDog - Monitor Messaging Queues
Summary:
To detect issues where a queue might not be getting drained and accumulating memories, this change adds counters to both `RWQueue` and `ReplicateQueue` types and export these counters to `fb303`.
Export is done periodically through the `watchdog` module. Currently, watchdog only logs the counters but does not take any action. In future (T98478475), `watchdog` will also crash if the queue size continues to increase.

In this diff, we only implement the queue counters and corresponding unit tests.

Reviewed By: saifhhasan

Differential Revision: D30256821

fbshipit-source-id: 86ec3e1e6ec7323cfecafc05e0ceb610ccc6ffa5