Use folly::F14NodeMap to improve cpu/memory perf in watchdog
Summary:
Introducing usage of `folly::F14NodeMap` and `folly::F14NodeSet` to improve performance

`std::unordered_map` or `std::unordered_set` are less efficient than folly map/set according to benchmark tests.

Ref for `F14NodeMap` and `F14NodeSet`: https://www.internalfb.com/intern/wiki/Hash_Tables/F14/

Ref:
- https://www.internalfb.com/intern/wiki/Users/grzybek/Performance/#thrift

The impact of this change may not be visible in the benchmark tests. But would like to verify if build for EBB build process has support for the library.

Reviewed By: xiangxu1121

Differential Revision: D26343869

fbshipit-source-id: bdd0750c7dc9baf5d650664a07a3170764755e48