Use Node Segment Label to generate prepend label based on SR policy config.
Summary:
In this change, we are supporting generation of prepend label (PL) using node segment label of the destination (or best node/area).
This is based on the mode as specified in the SR policy configuration and does not require PL generation based on the prepend label ranges.

PL is used to update the prepend label attribute in the best prefix Entry of RibUnicastEntry that is created for route distribution.

Reviewed By: cfdream

Differential Revision: D30403176

fbshipit-source-id: 61421994b8f10e65cd8fec9032f42ae276b02d6b