Increase the init RSS memory size
Summary:
It has higher RSS memory when the gtest process is running.
Increasing the first query limit.

NOTE: resident set size (RSS) is the portion of memory occupied by a process that is held in main memory (RAM). The rest of the occupied memory exists in the swap space or file system, either because some parts of the occupied memory were paged out, or because some parts of the executable were never loaded.

{F664158161}

This could due to the increasing of code, dynamically loaded lib and so on.

Reviewed By: TangoRoxy

Differential Revision: D30974094

fbshipit-source-id: abec707fb7023e91ba6bb12069cdfd5ecfa8badb