Fix MultipleAreaKeyExpiry testlet inside KvStoreClientInternalTest
Summary:
Looks like we have some issue in OSS environment when talking to thrift server via client-call.

For this `KvStoreClientInternalTest`, as the name suggests, there is no necessary to test the "EXTERNAL" API to check key existence.

As the test has used `getKey()` to validate key existence, we don't have to do repeated validation.

P.S. the `dumpAllWithThriftClientFromMultiple` will be tested individually with `KvStoreUtilTest`. Will handle it there separately.

Reviewed By: yi-xian

Differential Revision: D31067027

fbshipit-source-id: 5ce92da56887363989855203311e11a07b8588d0