(3/N): Split prepend_label_allocator and config out of giant openr:common lib
Summary:
As we want to simplify the dependencies for openr modules, I found that it is a pretty bad practice to make everything into one giant lib, for example, `openr/common:common`. Starting from this diff, we will try to simplify this common lib into small pieces.

One fundamental goal is to split config + openr if config out of common util lib.

Reviewed By: TangoRoxy

Differential Revision: D32178629

fbshipit-source-id: 27e24b1e697c4922b8f9fe5d165480cab164bcca