(4/N): Split Util.h/.cpp out of giant openr:common lib
Summary:
This is the continuous efforts to lose weight for the giant openr:common lib, which is included everywhere.

Some highlights:
- Make `Util.h/Util.cpp` a separate lib just as `NetworkUtil.h/Network.cpp` etc.
- One src code piece should just belong to one lib instead of multiple. `Constants.h/Constants.cpp` ONLY belongs to `openr/common:Constants` instead of `network_util`, `flags`, etc.;
- Clean up lots of unnecessary including of `openr/common/Util.h` to simplify lib dependency. This will have less chance of circular dependency in the future;
- Remove duplicate including of header files, like `NetlinkFibHandler.h`, `Watchdog.h`, etc.;

NOTE: found lots and lots of places we are including the header file without using it, maybe due to code deprecation/migration. Use auto-deps and buck build to make sure the updated one works.

Reviewed By: TangoRoxy

Differential Revision: D32198745

fbshipit-source-id: ab2e7f9d4710e4d23b96037976edd67cc8d33150