default to main branch
Summary:
Most referenced repos have moved from master to main as default branch, so flip the default in getdeps.

As well as tidying things up, this means more manifests will be able to use the pinned git commit hash for a dependency that shipit generates (e.g. build/deps/github_hashes/facebook/folly-rev.txt). This is because a "rev = " setting in the manifest is treated as a dynamic override in GitFetcher.__init__ and this diff reduces the number of such overrides.

Reviewed By: StanislavGlebik

Differential Revision: D32567564

fbshipit-source-id: 3f1b734eb1bb0f5dbafb01ceb935a3d51b8dcc9a