Migrate py2 un/framed transports to header
Summary:
Header transport is supported by all known thrift server implementations. Un/framed transports do not provide the minimum level of telemetry required and are super-deprecated (as opposed to header which is just regular deprecated).
Unlike with rocket upgrade, there's no negotiation because we don't know of any thrift servers to add it to that don't already support the header transport.

Reviewed By: cooperlees

Differential Revision: D33244874

fbshipit-source-id: be1635be52ad864f72d50cfa4a862dd98ce7d27a