(10/N): Remove KvStore dependency on general OpenrConfig
Summary:
To remove `KvStore`'s dependency on general Open/R (routing) config, we created a separate `thrift::KvStoreConfig` defined inside `if/KvStore.thrift`.

This will help with the dependency.

NOTE: this is NOT the final change as `thrift::KvstoreConfig`(not capital `s`) is still defined inside `if/OpenrConfig.thrift`. Will migrate towards that in future diffs for easy reviewing purpose.

Reviewed By: cfdream

Differential Revision: D33638460

fbshipit-source-id: 6aae3f15ffc344abe334ecc8e57ab6ab726f2e4a