(Bug fix 1/2) - Fix data retrieving of enableFloodOptimization flag
Summary:
With the usage of get_FIELD() API from thrift generated method, it is very dangerous to use it for `optional` field.

The issue here is that we set `enable_flood_optimization` field to be `false`. However, the get_XXX() API call will NOT retrieve the actual value of this field, but the **ADDRESS** of the field!

I use the following instrumental code:
{P478940558}

to validate the output:
```
Feb 07 21:07:15 fsw001.p006.f01.atn6 openr[519]: I0207 21:07:15.762806   519 Spark.cpp:341]
[DEBUG] config->getKvStoreConfig().get_enable_flood_optimization(): 0x7fd2937664d1,
getKvStoreConfig().enable_flood_optimization_ref().value_or(false): true
```

Basically, in Spark.cpp's constructor, we implicitly type-casting a non-zero address to a boolean variable `enableFloodOptimization_`. Since the address will never be 0x0, the value will always be set as TRUE regardless of what it is actually set.

Reviewed By: GirasoleY, cfdream

Differential Revision: D34064569

fbshipit-source-id: d55f01b88a85cfa08c9ee6188b8cced57c0fd8f1