Make lookup of LinkMonitor adjacency for an area efficient
Summary:
LinkMonitor adjacencies are keyed by remote node name and local interface name.

**Problem**
When a user needs to iterate all adjacencies for a given area, the above storage mechanism leads to unnecessary traversal of (and having to skip) adjacencies that belong to other areas.

**Solution**
To remedy this, we are changing how the adjacencies are stored to use a two level unordered_map. The first map's key is area name and the value is an unordered_map of all adjacencies with in the area.

The area adjacency is the original unordered_map keyed by node name and interface name and value as Adjacency itself (though it'll now only contains adjacencies that belong a specific area).

Reviewed By: TangoRoxy

Differential Revision: D34382415

fbshipit-source-id: eb8b81b4824be5e29f7b8583fb504872776358e1