(19/N) Make KvStoreServiceHandler and KvStoreWrapper template-based class
Summary:
This is the milestone diff to validate that we can talk in UT environment with either `ClientType` for KvStore synchronization.

# High-Level View:

```
KvStore - template class

      |

KvStoreServiceHandler - template class

      |

KvStoreWrapper - template class (ONLY FOR UT)
```
# Extensibility
Now we support any kinds of `ClientType` as long as we can provide the SAME `ClientType` KvStoreServiceHandler from thrift server's perspective. This makes easy to expand. Currently we provide:
- `thrift::KvStoreServiceAsyncClient` -> service defined in `KvStore.thrift`
- `thrift::OpenrCtrlCppAsyncClient`   -> service defined in `OpenrCtrlCpp.thrift`

Reviewed By: GirasoleY

Differential Revision: D34350035

fbshipit-source-id: 4bac054dbea87a2cb4858e340f2c10c406f989ea