Fix UT of subscribeAndGetKvStoreFilteredWithoutValue test
Summary:
This test doesn't check the value since we are expecting no value received as the test name suggested.

In test running, seems we can receive multiple updates:
- ttlVersion = 1, actually this is the initial snapshot of the key;
- ttlVersion = 2, the key we have injected with changes;

Fix to ignore the initial snapshot as we are NOT interested in that. Test purpose is making sure key update can be received.

Reviewed By: TangoRoxy

Differential Revision: D34531370

fbshipit-source-id: 3a125e7bd266c7bb3648765325b15e54625bc312