remove NEIGHBOR_DISCOVERY_ERROR
Summary:
LinkMonitor behavior is the same upon receiving `NEIGHBOR_DISCOVERY_ERROR` or `NEIGHBOR_DISCOVERY`. There's no need to introduce `NEIGHBOR_DISCOVERY_ERROR` state to confuse developer and users.

We consider `NEIGHBOR_DISCOVERY` the signals best effort to discover a list of neighbors. This list of established neighbors serve as the GR neighbors that a restarting node has to converge with.

Reviewed By: ashutoshgrewal, xiangxu1121

Differential Revision: D34809805

fbshipit-source-id: abbc2ca5e5f2f9bb41e54f6a27baaeb57809544b