Add zlib as an explicit dependency of airstore
Summary:
This adds zlib as an explicit dependency of airstore, so that it can be built & linked statically into the airstore client.

Technically, another dynamic dependency of airstore does itself have a dynamic dependency on zlib, but at this point, I think it's safe to say that is not something we can do much about, as that's likely a platform009 dependency.

Reviewed By: mackorone

Differential Revision: D35410511

fbshipit-source-id: 2ebfed431707d2ce73cfbf301c038c2da8fd30d9