Allow to use local EventBase from EventBaseManager with dumpAllWithThriftClientFromMultiple
Summary: Previously it was required to spawn or use detached EventBase with clients (otherwise it deadlocks) and to take care about destroying clients within this EventBase. After this change one can use EventBase from EventBaseManager. It is actually an EventBase of the current thread, so there is no need to spawn new one or to do any workaround to destroy clients correctly.

Reviewed By: xiangxu1121

Differential Revision: D35756953

fbshipit-source-id: ed0a9a43c30bd33bd37331ef2630f6ae0a0515df