Fix PrefixManager BGP route prioritization
Summary:
During the initialization, prefix manager would load the originated prefix and may wait for the BGP prefixes. The originated prefix should always be prioritized over BGP prefixes if `preferOpenrOriginatedRoutes_` is set. Currently, we don't have the guard over this prioritization. If the same prefixes are received from BGP after being built from the config during the initialization, we would overwrite the setting of the originated prefixes.

For instance, if the originated prefix should be programmed to fib, and the same prefix is received from BGP, the programmed prefix will be withdraw because the default behavior of a received BGP prefix is not programmed to fib. This violates the prioritization order.

To avoid so, we should check if

1. `preferOpenrOriginatedRoutes_` is true
2. the received BGP prefix is already built from the config

if both are true, we should not change the existing setting. This diff is addressing this issue and it includes the unit test `LocalRouteWinsOverBgpRoutesTest` to verify the result.

Reviewed By: ashutoshgrewal

Differential Revision: D36731949

fbshipit-source-id: 0708a144a27f282c16a5925746607d20eefd2654