Fix flaky test KvStoreTestFixture.PeerSyncTtlExpiry
Summary:
The test `KvStoreTestFixture.PeerSyncTtlExpiry` assumes that the KvStores will be in sync in 100ms, which is not necessarily true today. For example, [openr CI build failure log](https://fburl.com/testinfra/2l2s59z4) shows the processing time = 229 ms
```
I0529 22:26:52.726836 435678 KvStore-inl.h:1761] [Area test_area_name] [Thrift Sync] Full-sync response received from: store0 with 1 key-vals and 0 missing keys. Incurred 1 key-value updates. Processing time: 229ms
```
Therefore, the test did not go through and it would time out. A temporary workaround is the make wait longer to ensure that the sync would finish.

In the long run, we need a more reliable way to test, if possible.

Notice that the fix should be based on the fix of the getKey API in D36756150.

Reviewed By: xiangxu1121

Differential Revision: D36756693

fbshipit-source-id: 1747caff9ebd72f3aed56f6382c85413c499cca3