Create Dispatcher Module
Summary:
This is a new module called `Dispatcher` that will act as a wrapper for ReplicateQueue. It will use the underlying ReplicateQueue to filter keys based off regexes passed from the subscribers of `Dispatcher`. The `Dispatcher` will run on its own thread and when activated will allow for communications between `KvStore` and the other modules.

Link to RFC: https://fb.quip.com/AtqfAakENGSW

Reviewed By: shih-hao-tseng, xiangxu1121

Differential Revision: D37216065

fbshipit-source-id: bf67e52fb237a4886f918d9f1b0e6e4e458d7cb7