filter out value that has version=0
Summary:
# Overview
Income value with version=0 is invalid. We should ignore it and not merge it into kvStore.
# Background
See appended task. However, we do not have valid input to repro the crash and we do NOT have the coredump file.
# Solution
Nevertheless, we should still handle this case to mark it invalid

Reviewed By: xiangxu1121

Differential Revision: D37317151

fbshipit-source-id: 1e66cbaa71305aee015691512530ccba4c190dba