adding linear topology for convergence benchmark
Summary:
# Overview
We add `LINEAR` topology, which is basically just connecting each peer in the vector with each other.
# Detail
Imagine a vector of kvstores. We generate a linear topology by:
1. having prev pointer points to index 0
2. having cur pointer points to index 1
3. prev and cur adds each other as peer
4. prev = cur
5. cur++
6. repeat 3-5 until exhaust vector
# Note
This is work part of understanding how kvstore convergence behaves. Furthermore, this is part of the setup cost, and should not count toward convergence time

# Facebook
Thoughts on using bi-directional link between peers vs. single direction?

Reviewed By: xiangxu1121

Differential Revision: D37539549

fbshipit-source-id: d55d01d0e1bb67d50ca2462e7289af3d0391a306