Dispatcher Integration
Summary: Integrating the `Dispatcher` into `Main.cpp` with the use of a knob. The knob will still be switched off for the current versions of Openr. Slight modification to `PrefixManager` to account for the case where `Dispatcher` knob is switched on. This diff only concatenates `Dispatcher` with `PrefixManager` and `Decision`, but not `OpenrCtrlHandler`.

Reviewed By: shih-hao-tseng

Differential Revision: D37699563

fbshipit-source-id: 277f6831f316a125ddd6d8e61bd2179d064c2f17