Moving DispatcherQueue Initialization Outside Dispatcher
Summary: Moving DispatcherQueue initialization outside of Dispatcher to prevent deadlock issue. The deadlock was being caused when the RQueue was supplied by Dispatcher to the internal modules. We needed Dispatcher to close all the queues explicitly first and then be stopped. Since we didn't break this process into two separate steps, the other modules were stuck in infinite loops while they waited on updates to the RQueue (since the queue hadn't been closed first) and Dispatcher was one of the last modules to be stopped. This change will also allow us to get performance metrics from watchdog.

Reviewed By: xiangxu1121

Differential Revision: D37858503

fbshipit-source-id: 4f278142481a214cdd58e14361f2fbd1266857ba