Enhance breeze lm links cmd to show soft-drain metric
Summary:
As titled. Now we don't have a way to show if a node is soft-drained or NOT without explicitly checking the node metric.

This diff adds the support for `breeze lm links` support for soft-drain.

NOTE: ideally, this should be added with `breeze decision adj`. However, the AdjacencyDatabase change is yet to hit PROD. Do it separately.

Reviewed By: TangoRoxy

Differential Revision: D39512774

fbshipit-source-id: f3b1fb141acb7ac0af08e9cc750ee9f2068c3bbc