Refine initial Spark neighbor discovery procedure
Summary:
D38604812 (https://github.com/facebook/openr/commit/f051dbb022ba8a9432096299596fccf10522a594) allowed Spark to publish ```NEIGHBOR_DISCOVERED``` signal to ```LinkMonitor``` after all known neighbors are ```ESTABLISHED```. The goal was to avoid unnecessarily waiting for a timer to expire even after all the known neighbors have been discovered.

However, publishing the ```NEIGHBOR_DISCOVERED``` signal when all known neighbors are  ```ESTABLISHED``` may be premature. This is so because ```Spark``` only knows about a neighbor when it receives a ```SparkHelloMsg``` from the neighboring node. This means it's possible that all known n-1 neighbors could ```ESTABLISH``` just before the nth (potential) neighbor was about to send it's first ```SparkHelloMsg```.

Thus, we introduce the concept of a min-max interval in ```Spark``` that controls when it can publish the ```NEIGHBOR_DISCOVERED``` signal to ```LinkMonitor```. The min interval is a lower bound that must elapse before considering publishing this signal while the max interval is an upper bound at which point the signal must be published. The exact behavior is as follows -

- At the expiry of the min interval, ```Spark``` may publish the signal if all it's known neighbors are ```ESTABLISHED```.

- If there are 1-or-more known neighbors that are still not ```ESTABLISHED```, ```Spark``` will continue to wait for them to ```ESTABLISH```. When these stragglers become ```ESTABLISHED```, ```Spark``` will publish the signal.

- However, this wait is upper bounded, by the max interval, at the expiry of which ```Spark``` will publish the signal irrespective of whether all known neighbors are  ```ESTABLISHED``` or not.

Reviewed By: xiangxu1121

Differential Revision: D39377686

fbshipit-source-id: 4af707705afd1a0c119f3af7763f87c4a79473b9