Fix mem leak when no local BGP peering enabled
Summary:
As titled. A mem leak instance inside Open/R.

`prefixMgrRouteUpdatesQueue` is the queue used from `PrefixManager` to local `BgpRib` if enabled. The reader queue is created ONLY when bgp peering is there.

https://www.internalfb.com/code/fbsource/[267b7d02941880809f1b918227fda9c59cf72bec]/fbcode/openr/Main.cpp?lines=265-271

NOTE: ideally we can check it out by using the watchdog queue count monitor. Unfortunately this queue is NOT added to the Watchdog. :(

Reviewed By: TangoRoxy

Differential Revision: D40925483

fbshipit-source-id: 5bc1b3196699313377c19b05558bc42d02104a65