Fix decision not using newly updated prefixEntry
Summary:
As titled. Didn't update the ribEntry with correct prefixEntry
This will fix the display for remote routes but not the ones in the same area. That will be done in a separate diff

Note: when a node in the area is drained, display is incorrect because we use prefixState rather RouteDB for metrics. Will address in next diff.

Reviewed By: xiangxu1121

Differential Revision: D40952824

fbshipit-source-id: 48194dd0c3a445868ebd97c245d3831e61bf4e70