set stream expiration to infinite
Summary:
As title.

The default stream expiration time is set to 1min. If client does not send any request within that time, the server would close the connection to prevent bad behaving client. For our use case, client needs to indefinitely subscribe to the server. So we set it to infinite such that the connection would not get closed.

Reviewed By: TangoRoxy

Differential Revision: D41280481

fbshipit-source-id: e7e50e0588694c21cc63b37082c6328bbee14eb0