Fix SparkFixture.ForcedInitializationTest
Summary:
SparkFixture.ForcedInitializationTest was flaky due to the following reasons:
1. The elapsed time was not measured right after the initialization event was received, which introduced some measurement error.
2. The test deliberately tries to verify that the elapsed time would hit the `maxNeighborDiscoveryInterval`. However, due to the issue 1, the measurement could be slightly over the limit and fail the condition.

For example, the `maxNeighborDiscoveryInterval` could be 4000 ms. However, because the `ElapsedTime` was measured after the initialization event is received and the control flow returned to the test thread, there would be some context switching overhead. Also, in the original test, the `ElapsedTime` was measured after some other logic, which introduced some futher measurement error and made it something like 4004 ms. If we assert that `ElapsedTime <= maxNeighborDiscoveryInterval`, the condition would fail.

This diff fixes the issues and makes the test stable.

Reviewed By: TangoRoxy

Differential Revision: D41588139

fbshipit-source-id: 96c323bed782e2db7fe5b0a38fdf5a238e7d04f4