Mitigate racing issue in SparkFixture.HubAndSpokeTopology
Summary:
SparkFixture.HubAndSpokeTopology is marked flaky and it could sometimes fail under stress-run. Looking at [the log](https://fburl.com/testinfra/sisk0gii) and we could see that the issue was
```
[ RUN      ] SparkFixture.HubAndSpokeTopology
fbcode/openr/spark/tests/SparkTest.cpp:1442: Failure
Value of: node3->getTotalNeighborCount() == 0
  Actual: false
Expected: true
```

The corresponding lines are at
https://fburl.com/code/s3bkx0ji. In the code, we will first notify LinkMonitor about the neighbor down event through `neighborDownWrapper(neighbor, ifName, neighborName);` then decrement the total neighbor count by `eraseSparkNeighbor(ifNeighbors, neighborName);`. There is a racing issue here: when the function `waitForEvents` captures the neighbor down event notification, the total neighbor count is not necessarily decremented yet (while the active neighbor count should be correctly updated already). Therefore, under stress run, we could hit the racing condition as the test thread is different from the spark thread.

Ideally, we should decrement the number then send out the notification, but it is not easy to do: we need to first remember the neighbor info so that we could craft the notification after the neighbor info is removed and the total neighbor count is decremented.

In the real environment, it would be very unlikely to hit this issue unless we have a thread aggressively monitoring the neighbor down event and insisting that the total neighbor count should be decremented upon the time when the event is sent. Given that the impact is not significant and the fix is too cumbersome, this diff mitigates the flaky test by just checking that the total neighbor count would go back to 0 in 5 ms (checking every ms).

Similar issue could happen for heartbeat timeout as well. So we also fix the related issues in this diff.

Reviewed By: ashutoshgrewal

Differential Revision: D41620497

fbshipit-source-id: a1b83ea82e73dfb81cfa29a82b627f4e8933d45e