Migrate openr audit to async client
Summary:
The route db validation in `openr_audit` tool still uses thrift-py based sync client. This diff migrates it to the thrift-python based async client so that we could deprecate the thrift-py utility function `get_routes_py`.

According to the following diffs, it seems that the tool is no longer in use and we just migrate to the async client without guaranteeing its functionality.
- D27278233 removed the tool already.
- D31722388 fixed the tool for turbo fabric, but turbo fabric is deprecated.

Reviewed By: TangoRoxy, xiangxu1121

Differential Revision: D41979496

fbshipit-source-id: a5eab658e157cd92d50c1eca1da363e46657d4db