Fix createFibClient retry issues
Summary:
When OpenR tries to connect to wedge agent and the wedge agent is not yet ready, OpenR would retry the connection establishment. Current logic could lead to crashing as shown in P590138913 by ashutoshgrewal. The error message shows that
```
pure virtual method called
```
at the following line
https://www.internalfb.com/code/fbsource/[8e358b3e9b1acb8298c8b6cb4f5c7d8898887aa5]/fbcode/openr/fib/Fib.cpp?lines=1065

Using fdb, we could find that the issue happens when calling `!socket->good()` where `socket` points to an object of the type `AsyncTransport` despite having its type `AsyncSocket*`. The reason was that the actual object was kept by an unique pointer `newSocket`, which eventually got sent into `client` through `channel`. After async connection, when the client failed to connect, it freed the actual socket object.

Now, `socket` still points to the original socket object and calling `socket` should have led to a SegFault. However, the `AsyncSocket` has delayed destruction, which prevented this happening. Therefore we see the pure virtual method error instead of a SegFault.

The issue here is the discrepancy between the `socket` pointer and the actual socket object used by the `client`. Revisiting the codes and we could see that the `socket` is only used by the `createFibClient` function. There is no need for us to pass the `socket` object around and use a potentially out-dated object as an indicator of whether we should reset `client`. We should just check if the channel of the client is good. If not, we reset the client. This diff does just that.

Reviewed By: ashutoshgrewal, xiangxu1121

Differential Revision: D42338143

fbshipit-source-id: eae121699638b6ef059e2a85bb30af2f979df48f